elNamed post

+++ .Rprofile
@@ -1,10 +1,12 @@
 fuse_site <- function() {
-  file.remove("index.html")
+  if (file.exists("index.html")) file.remove("index.html")
   litedown::fuse_site()
   litedown::fuse_site("index.html")
   browseURL("index.html")
 }
 
 newpost <- function() {
-  file.copy("posts/_template.Rmd", tempfile("aaa-", "posts", ".Rmd"))
+  f <- tempfile("aaa-", "posts", ".Rmd")
+  file.copy("posts/_template.Rmd", f)
+  browseURL(f)
 }
